pyproject.toml
[project]
name = "bitrix-lead-analyzer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "pocketsphinx>=5.0.4",
    "pydantic>=2.11.7",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "schedule>=1.2.2",
    "speechrecognition>=3.14.3",
    "uvicorn>=0.35.0",
]
